QuickAdminPanel file update

+++ database/migrations/2020_11_18_000013_create_user_user_alert_pivot_table.php
@@ -9,9 +9,9 @@ class CreateUserUserAlertPivotTable extends Migration
     public function up()
     {
         Schema::create('user_user_alert', function (Blueprint $table) {
-            $table->unsignedInteger('user_alert_id');
+            $table->unsignedBigInteger('user_alert_id');
             $table->foreign('user_alert_id', 'user_alert_id_fk_2573353')->references('id')->on('user_alerts')->onDelete('cascade');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id', 'user_id_fk_2573353')->references('id')->on('users')->onDelete('cascade');
             $table->boolean('read')->default(0);
         });
