+++ database/migrations/2020_11_18_000015_create_car_carmodel_pivot_table.php
@@ -9,9 +9,9 @@ class CreateCarCarmodelPivotTable extends Migration
     public function up()
     {
         Schema::create('car_carmodel', function (Blueprint $table) {
-            $table->unsignedInteger('carmodel_id');
+            $table->unsignedBigInteger('carmodel_id');
             $table->foreign('carmodel_id', 'carmodel_id_fk_2578973')->references('id')->on('carmodels')->onDelete('cascade');
-            $table->unsignedInteger('car_id');
+            $table->unsignedBigInteger('car_id');
             $table->foreign('car_id', 'car_id_fk_2578973')->references('id')->on('cars')->onDelete('cascade');
         });
     }
