+++ database/migrations/2020_11_18_000018_add_relationship_fields_to_users_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('team_id')->nullable();
+            $table->unsignedBigInteger('team_id')->nullable();
             $table->foreign('team_id', 'team_fk_2472572')->references('id')->on('teams');
         });
     }
